Replace some defines with ptrdiff_t -- a portable integral type, to get rid of warnings and increase portability.

+++ trunk/source/game/g_local.h
@@ -513,13 +513,13 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define	FOFS(x) (int)&(((edict_t *)0)->x)
-#define	STOFS(x) (int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (int)&(((gclient_t *)0)->x)
+#define FOFS(x)(ptrdiff_t)&(((edict_t *)0)->x)
+#define STOFS(x)(ptrdiff_t)&(((spawn_temp_t *)0)->x)
+#define LLOFS(x)(ptrdiff_t)&(((level_locals_t *)0)->x)
+#define CLOFS(x)(ptrdiff_t)&(((gclient_t *)0)->x)
 
-#define random()	((rand () & 0x7fff) / ((float)0x7fff))
-#define crandom()	(2.0 * (random() - 0.5))
+#define random() ((rand() & 0x7fff) / ((float)0x7fff))
+#define crandom() (2.0 * (random() - 0.5))
 
 extern	cvar_t	*maxentities;
 extern	cvar_t	*deathmatch;
